SDC-7944: Ensure that topic exists before running Kafka origin tests by firstly generating records
This is a bit of a workaround change - The new Multitopic Kafka origin
frequently times out because the pipeline runs before the actual topic
exists. Now there is a bug somewhere in Kafka's own NetworkClient that
doesn't honor the timeout and hence we end up waiting for a while which
eventually times out the test. This particular trouble seems to be fixed
in Kafka 0.11 and 1.0, but sadly persist on 0.9 and 0.10.

Since our tests are not explicitly creating topics, but rather depend on
the Kafka auto-topic creation concept, I moved the message production
before we start pipeline - this way we know that the topic exists (and
have messages) before our pipeline runs which then works properly.

Change-Id: If6a781172e6500667ab302c37173f8624e9ac82c
Reviewed-on: https://review.streamsets.net/11677
Reviewed-by: Kirti Velankar <kirti@streamsets.com>