SDC-10319: datacollector.test_classpath.test_mapr_classpath failing
Skipping testing MapR as it currently have duplicates all over - our
linking script takes input list of directories and links all jars from
them. It seems that mapr doesn't do a good job in synchronizing
dependencies, so we end up with bunch of duplicates. The best solution
would be to rewrite our linking script, but that is an undertaking on
it's own so I will simply disable this test for now.

Change-Id: I8fd75b97b680ca0a32cdc5860e1257d39c1fb165
Reviewed-on: https://review.streamsets.net/17870
Reviewed-by: Junko Urata <junko@streamsets.com>