SDC-11350: New JDBC tests are failing on some databases
The trouble is that the test need to properly check offset which is
slightly difficult as each database have slightly different table stored
in the offset (SDC.$TABLE, dbo.$TABLE, public.$TABLE). Rather then
encoding those differences now and then in the future for each every db,
this patch limits execution for MySQL only. Since this is internal
behavior that doesn't depend on the database, it seems sufficient thing
to do.

Change-Id: I892ecb3cea13f7b6d555d2607911df8335cab90b
Reviewed-on: https://review.streamsets.net/c/21228
Reviewed-by: Xavier Baqués <xavi@streamsets.com>