SDC-12383: Update exported test pipelines for changed static validation
We have recently updated our static validation code to explicitly
validate if someone enters invalid field path into a field that is
annotated to expect a field path - which is something that we were
historically doing only during runtime. Bunch of the exported old
pipelines however have bogus values in those fields and thus are no
longer valid. This patch updates those invalid values so that we can
continue running the upgrade tests.

Change-Id: I6e208b497bbad700e7c5e4c34b494fac24f0ce53
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/1148
Reviewed-by: Xavier Baques Arcusa <xavi@streamsets.com>