SDC-11216. test_rabbitmq_stages.test_rabbitmq_rabbitmq_consumer is failing
The issue came from a change in the 'pika' library, employed by us to connect to
the AMQP server and send the messages to be consumed by the stage. The
`Channel.basic_publish()` method does not return any value anymore, but we
expected a confirmation was returned.

See for more details
https://github.com/pika/pika/blob/cf954e1b68e5a5f91cd5605d8ed6aaa3dc2ed308/pika/channel.py#L403

Change-Id: I3ad2081704c32917533915682ccf26f3ceeb04fc
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/1309
Reviewed-by: Xavier Baques Arcusa <xavi@streamsets.com>