SDC-12841. 	Integration test for SDC-12835
Configuration:
pipeline retry is set to True
onErrorRecord action is set to STOP-PIPELINE

CurrentBehavior:
When the Kafka Destination is stopped due to a stage Exception whilst in RUNNING state(say network connectivity) OR never started(say due to an invalid Kafka configuration element - compression.type = invalid.value), then, currently, the pipeline is stopped and never retried.

Explanation for current behavior:
Even though the pipeline is set to retry, the way that the state transtion from error to retry is handled is that, the code checks whether the onErrorRecord action for any stage is set to STOP-PIPELINE or not.
If it is, then the state transtion logic doesnt retry the pipeline again.
- This is not the expected behavior, since a stageException or StartError is not the same as a record error, and as such, should be treated differently

NewBehavior:
Stop the pipeline and do not retry only if it is indeed a record error
For other cases, like stageException or startError, conform to the retry configuration

STF Setup:
Connect a random generator to KafkaProducer destination
Induce a start error by setting an invalid Kafka configuration element (compression.type = invalid.value)
Verify that, with the fix in place in 3.13, the pipeline is retried after raising a START_ERROR
In versions < 3.13, this test should fail. That is expected behavior

Change-Id: I531058d0439e654449a8117a793042d33ea9ea59
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/1340
Reviewed-by: Dima Spivak <dima@streamsets.com>