SDC-13450: Revisit test_ftp_destination_create_path_pathrelative_finisher
The test can't possibly test what it aims because the server we're using
is configured to always chroot given user. Thus whether we are using
"user dir" as root or not doesn't matter as that is exactly the same.

Even if we would reconfigure the default environment, the test would
still need updates to ensure the difference in file names that we're
reading. Thus I feel that for now, it's simply easier to drop the test
as it's not adding new value. The test test_ftp_destination already
covers upload of files.

Change-Id: Idd60e8823fd2fe26aabe4128e28f2015812785ea
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/1510
Reviewed-by: Virag <virag@streamsets.com>