SDC-10427: test_jdbc_tee_processor_multi_ops[False] failure on SQL Server
The core reason of why the test fails on SQL Server is that the test
isn't written to be compatible with SQL Server - the underlying table
isn't properly created with identity column that is required for the
functionality to work.

While making the test passed, I've noticed major behavioral changes that
SQL Server does when comparing to let's say MySQL or PostgreSQL - for
example when you do delete operation, the keys aren't properly fetched.

Interestingly enough, other Tee tests are actually being skipped for SQL
server, so I made them to skipp all and will follow up with docs to say
that both SQL Server and Oracle aren't really supported at this time.

Change-Id: I1db4db2bf1ab1f1a937126bba6ac6d40edcfdf0d
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/1531
Reviewed-by: Virag <virag@streamsets.com>