SDC-13769: Update exported pipelines for old upgrade tests
This is fixing pipeline that was missed in SDC-12383. Original commit
message follows.

We have recently updated our static validation code to explicitly
validate if someone enters invalid field path into a field that is
annotated to expect a field path - which is something that we were
historically doing only during runtime. Bunch of the exported old
pipelines however have bogus values in those fields and thus are no
longer valid. This patch updates those invalid values so that we can
continue running the upgrade tests.

Change-Id: I2c86d2351c269bb445672e88fe747eb52c62bb57
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/1620
Reviewed-by: Alex Sánchez <alex.sanchez@streamsets.com>