SDC-13771: Drop test test_sql_server_change_tracking_on_unknown_type_action
The test was added during time when we haven't supported DATETIMEOFFSET
type, which was however later on added via SDC-10125. I've tried
changing the type to something else, but sadly it seems that the JDBC
driver is converting various custom types into binary type (when asking
over JDBC) which makes it impossible to test the ON_UNKNOWN_TYPE
behavior (since there is no unknown type reported to us any more).

Change-Id: I1494e4ca709792e3f887c7a6f9b7bad24c499c4d
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/1621
Reviewed-by: Alex Sánchez <alex.sanchez@streamsets.com>