SDC-13773: Drop test test_sql_server_cdc_with_zero_initial_offset
As far as I can tell the test was failing ever since it was introduced
and as far as I can tell it's pretty much copy&paste of another test -
test_sql_server_cdc_with_nonempty_initial_offset with only few changes
here and there. The trouble is that offset '0' per doc and per the code
isn't special in any way. It marks the start point so I don't think that
the test have right logic and thus I would recommend to simply drop it.

Change-Id: I3e3ce8537a6c7f1bee490f8cdf6251c025bd1321
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/1622
Reviewed-by: Alex Sánchez <alex.sanchez@streamsets.com>