SDC-12823: stage.test_aws_s3_origin.test_invalid_configs_diff_bucket_same_prefix is failing in bifocals
In order for us to generate the error S3_SPOOLDIR_12, given configured
bucket must not exists - but buckets have global namespace. So what
happened is that someone somewhere on the internet created the bucket
'post-process-bucket' and thus our test started failing. This patch
updates the bucket name to some random string do decrease the chance of
that happening again.

Change-Id: I42a21bc963e8167ceef65c8fb38ae9171f0ec793
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/1640
Reviewed-by: Karthik Iyer <karthik@streamsets.com>