SDC-13869: Drop test test_sql_server_cdc_on_unknown_type_action
Now the test fails in a support method that when fixed fails the same
way as described in SDC-13771. Thus I suggest to solve it the same way:

The test was added during time when we haven't supported DATETIMEOFFSET
type, which was however later on added via SDC-10125. I've tried
changing the type to something else, but sadly it seems that the JDBC
driver is converting various custom types into binary type (when asking
over JDBC) which makes it impossible to test the ON_UNKNOWN_TYPE
behavior (since there is no unknown type reported to us any more).

Change-Id: Ibd7c09726772b1c7ead3cecac09f7d9150765adc
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/1650
Reviewed-by: Karthik Iyer <karthik@streamsets.com>