SDC-12732: pipeline.test_kafka_schema_registry.test_kafka_consumer_key_capture_modes is failing on Jenkins
Based on prior research this is bug inside Kafka itself that manifests
only in the test and not in the product. Since Kafka 0.9 is really very
old, I don't see us troubleshooting as good use of our time so I will
skip the tests inside that version for now. And in parallel work on a
plan to remove Kafka 0.9 support altogether.

Change-Id: I613114544e7e8b7742504758124652ddf10b2370
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/1655
Reviewed-by: Karthik Iyer <karthik@streamsets.com>