SDC-13887: Test test_kafka_xml_record_delimiter_element_cluster is failing
Seems that the test fails assertion in every CDH distribution that we're
testing with. The history of the test seems to imply that it work
differently at some point in the past, but it's hard for me to triage
when since the XML parsing is done by the underlying Hadoop version and
not by us. Hence I will just update the expectations for now, so that we
can clearly document what is expected for anyone who will ask.

Change-Id: I9d808d30055a6442ec7e65c5d7066c88c1dd86c1
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/1658
Reviewed-by: Karthik Iyer <karthik@streamsets.com>