SDC-12946: Allow S3 processors to specify virtual or path addressing model
Writing test for this improvement is kind of difficult as the new config
is propagated down to the AWS SDK and there is no noticable change in
behavior that we can easily detect. The only real way how to test what
the AWS is stating it will do would be wireshark and listen to the way
the SDK communicates with AWS servers.

That seems as both overkill as well as not our job to do (since that
would be really testing just AWS library and that it does what they
state it does). Thus I chose a different approach, just setting the
config and validating that everything still works with both True/False.

Change-Id: If5d2caf35c443e1ab2b1b8b6906494ccfbb8cc19
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/1752
Reviewed-by: Alex Sánchez <alex.sanchez@streamsets.com>