SDC-14158: Use unique prefix for stage/configuration/test_sftp_ftp_ftps_client_origin.py::test_file_name_pattern_mode
I'm having troubles to repro the failure exactly as it's seen in
Bifocals, but when I upload 'random' files to the SFTP server, I can
reproduce similar errors. Thus I'm assuming that some tests are not
properly cleaning up which breaks this test. As a result I use a prefix
here to avoid random colision.

Change-Id: I08dbf8897a79c80866ac1a3de60ba5d9ec883997
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/1809
Reviewed-by: Karthik Iyer <karthik@streamsets.com>