SDC-14161: Add additional logging into stage.test_directory_origin.test_directory_origin_multiple_batches_no_initial_file
I see the test intermittently failing on Jenkins - it's not clear to me
if it's flaky or the comparison fails for a different reason. Since it's
'randomized' test I'm hoping that additional logging will help me answer
that question.

I wished I would be able to repro the failure on my box - tried 100
times in a loop and did not failed a single time :-/

Change-Id: I574855b54fc3a64df3b6b697bf2e78c67eec31e6
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/1812
Reviewed-by: Karthik Iyer <karthik@streamsets.com>