SDC-14340: stage.configuration.test_field_splitter_processor.test_field_to_split is failing in upgrade tests
In the earlier versions of SDC the field data format wasn't filled at
all - e.g. user had to explicitly configure something. At some point we
have added default value which this test was depending on. But since we
need to run it in upgrade fashion, we have to set the data format
explicitly.

Change-Id: I753bf052742b3451dfd2d69da73d060db99ea742
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/1919
Reviewed-by: Karthik Iyer <karthik@streamsets.com>