SDC-14437: stage.configuration.test_field_splitter_processor.test_field_to_split is failing on certain SDC versions
The reason why it's failing on some versions and not others is a bug
with default value that we have inside the core engine - covered by
SDC-12751.

However the test itself is asserting data based on a default value of a
config field which doesn't seem right - we can change that value any
time and thus the test should explicitly specify it.

Change-Id: I8c19a1b7d9af7057979e9bd5ff84dc9612100aa0
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/1967
Reviewed-by: Alex Sánchez <alex.sanchez@streamsets.com>