SDC-14476: Temporarily log the RAW snapshot data for Kafka tests
I'm not able to repro the issues tracked by SDC-14423 locally and even
on Jenkins it happens very occasionally. Thus I'm trying to instrument
the code a bit better and put some additional dumps to help me
understand what is going on. The going theory of mine is that the
machine is a bit overloaded on Jenkins and the first batch is simply
empty.

Change-Id: I9d34edcb848fcf412b96e4c22b36777607365080
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/1984
Reviewed-by: Karthik Iyer <karthik@streamsets.com>