SDC-14469. Google Storage Destination Tests are failing with incorrect error record being thrown
As a part of SDC-14374, STFs for testing google cloud storage destination error record handling was added to stages/configuration/test_google_cloud_storage_destination.py
So, this is not required in stage/test_gcp_stages.py anymore. Hence, removing it.

Also, since we are now using proper error record handling, the error message first reflects an error generated by the stage
So, changing the check for CONTAINER_0001 to GCS_09

Change-Id: I58bf70c92d00c6e14d4da4a281e540240c6063f8
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/1985
Reviewed-by: Matt Bayley <bayley@streamsets.com>