SDC-14887. stage.test_jdbc_stages.test_jdbc_consumer_non_incremental_mode failing for some environments
The test failed for MySQL and MemSQL environments when inserting the input data via
SQLAlchemy. These environments has a especial treatment for value 0 in primary
keys. Instead of a raw value, it is used to request a new incremental value generated by
the database. That provoked duplicate keys for value 1.

Change-Id: Ib33ea8cd93fe3990ea5365ceb22dd64683e54f54
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/2162
Reviewed-by: Alex Sánchez <alex.sanchez@streamsets.com>