SDC-15197. Increase snapshot capture time for s3 tests
This patch increases the snapshot time for tests in s3 origin. Recently we have seen the object spooling takes a bit of time to get back the results and default snapshot time of 30 seconds is not enought.

For future: We use ListObjects api in SDC today, but s3 recomments ListObjectsV2 Api, created SDC-15198 to track the work that might be needed

Testing:
Ran test_s3_origin_data_formats test which is failing on older versions

Change-Id: Ifb3d5feb1808ab6d9813f1600bdf3433df447e6b
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/2276
Reviewed-by: Eddie Park <eddie@streamsets.com>