SDC-15343: stage.test_jdbc_stages.test_jdbc_producer_no_implicit_mapping is failing for Oracle environments
Skipping in Oracle when not using explicit mapping - since that leads to
troubles with column cases. Since the test isn't specific to Oracle I
feel that for test readability is to simply skip Oracle in that case.

I've also moved the asserts around since that also improves test
readability - seeing more or less error records is much more relevant
then seeing what is in the database. Especially when something
unexpected happens.

Change-Id: I945e405c3965ca7143dd5970dabc85f63ccbe2aa
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/2331
Reviewed-by: Alex Sánchez <alex.sanchez@streamsets.com>