SDC-15422. test_oracle_cdc_client_preview_and_run[buffer_locally=False] fails with timeout
This patch fixes the case where buffer_locally=False fails in the above test. The main crux it timesout it because in ORACLE CDC there is a background generation thread which connects and read data from log miner. The oracle driver supposedly behaves in a way that even if the thread is shutdown from the main destroy thread, the connection is not closed until timeout kicks in from server. This patch generally follows what other tests are doing by force killing the pipeline.
The above scenario is mainly prelevant in ORacle_11_2_0 environment, ORacle_19 automation is fine
For later, we probably need to change the timeout in automation so generally automation completes in acceptable times and pipelines are not force killed.

Change-Id: I18c4eca7b63acf706632778410729c2ad7739a14
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/2385
Reviewed-by: Eddie Park <eddie@streamsets.com>