SDC-15714. Test if the path is consistent in events of LocalFS
We want to test that:
1. The filepath attribute in events generated by the Local FS, ADLS, HDFS
destinations contain only the path and doesn't contain schema or host.
To test this we create a file which by the corresponding FS destination
will be considered as an old file. That forces SDC to create
an event. Due to a bug the filepath attribute of the even would contain
a schema. With the fix the attribute will contain only the path.
The FS destination gets its input from Dev Raw Data Source
The event will be sent to the Trash destination.

2. The implementation of the WHOLE_FILE mode still works as before,
because it wasn't broken. So we also want to make sure the fix doesn't
break it. Thus we test the same set of the FS destinations using the same
approach of creating "old" files but we feed records from the Directory
source, that allows us to enable the WHOLE_FILE mode.

Change-Id: Ie12a5455e8c5a6775c61abf373c19903bcb91bd0
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/2525
Reviewed-by: Santhosh Kumar Manavasi Lakshminarayanan <santhosh@streamsets.com>
Reviewed-by: Dima Spivak <dima@streamsets.com>