SDC-15847: Move JDBC Producer tests to it's own file
* test_jdbc_producer_insert
Removed, it's smoke test exercised by several other tests.

* test_jdbc_producer_insert_type_err
Moved to it's own file.

* test_jdbc_producer_no_implicit_mapping
Moved to it's own file.

* test_jdbc_producer_insert_multiple_types
Removed, standard tests are testing all types better.

* test_mssql_producer_bigdecimal
Moved to it's own file.

* test_jdbc_producer_coerced_insert
Removed, standard tests are testing all types better.

* test_jdbc_producer_delete
Moved to it's own file.

* test_jdbc_producer_update
Moved to it's own file.

* test_jdbc_producer_multirow_with_duplicates
Moved to it's own file.

* test_jdbc_producer_multitable
Moved to it's own file.

* test_jdbc_producer_multischema
Moved to it's own file.

* test_jdbc_producer_multischema_multitable
Moved to it's own file.

* test_jdbc_producer_ordering
Moved to it's own file.

* test_jdbc_producer_oracle_data_errors
Moved to it's own file.

* test_error_handling_when_there_is_no_primary_key
Moved to it's own file.

Change-Id: I518164412c414885f3a04e1720be1b74db727528
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/2566
Reviewed-by: Alex Sánchez <alex.sanchez@streamsets.com>