SDC-15861. Test the + character gets encoded correctly for ES5/6
We want to test that we can search in documents of document types
containing the '+' character belonging to indices containing the '+'
character. So we will create an index with the '+' character and a
document type with the '+' character. Then we will search for fields in
all in the index. If the '+' character is accepted then we will get
exactly the same document we created before since the newly created
index contains only it.

This same test should pass for all Elasticsearch versions.

Change-Id: Id79304a836b6505cd23a62d493700e610432a663
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/2576
Reviewed-by: Santhosh Kumar Manavasi Lakshminarayanan <santhosh@streamsets.com>