SDC-15920. Test query validation in ElasticSearch origin
We will test a valid query containing an additional field apart from the
query field. In the validate API, only the query field is allowed,
though in the search API other fields may be used, e.g. the sort field.
Here we want to test that if a valid query contains additional fields
the origin doesn't break.

We also want to test that the validation fails if a query is invalid or
a query JSON is not a valid object itself.

Change-Id: Icc08b328085bcf954782d3bb35d4d5559c602b42
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/2608
Reviewed-by: Martin Balzamo <martin@streamsets.com>
Reviewed-by: Santhosh Kumar Manavasi Lakshminarayanan <santhosh@streamsets.com>
Tested-by: <streamsets-ci@streamsets.com>