SDC-15945: stage/standard/test_jdbc_producer_destination.py::test_multiple_batches is flaky
This is actually my bad, I knew when writing the test that the pipeline
can be fast and generate more then expected records. That is why the
code was checking actual amount of ongoing records and validating that
it's at least expected count. But then the comparison incorrectly used
the wrong variable.

Change-Id: Ieec0ee175db357a37c1d63e2b1348eb193a9b0e1
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/2626
Reviewed-by: Alex Sánchez <alex.sanchez@streamsets.com>