SDC-15953: stage.configuration.test_kafka_consumer_origin.test_allow_extra_columns is flaky
The tests were quite frequently failing with "operation only permitted
on running pipeline". Looking into the test, the pipeline is created as
a batch pipeline - yet we have explicit wait in one arm of the test.

So I've refactored the test to know that it's batch pipeline and always
wait for the pipeline to finish.

I've rerun the test locally 20 times and always passed.

Change-Id: I697bbb4d089a8fbccca4d3703d4e9a38c1431c62
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/2634
Reviewed-by: Alex Sánchez <alex.sanchez@streamsets.com>