SDC-16002: stage.test_postgresql_metadata_processor.test_unsupported_type_in_precreated_table is failing
The problem is that the test depended on the fact that the processor
haven't properly support SHORT before (that mapped to smallint). I've
however added this and all other types recently and hence the test on
it's own no longer makes sense. I've at least rewrote it to do something
useful and that is to ensure that we properly create error records when
the types are a mismatch.

Change-Id: I12a4a471deba713a2249ad01265504d46ccfa3c8
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/2683
Reviewed-by: Alex Sánchez <alex.sanchez@streamsets.com>