SDC-16003: stage.standard.test_jdbc_producer_destination.test_data_types_postgresql from ZONED_DATETIME to text types are failing
This is a fallout from my change that added proper support for
ZONED_DATETIME - since we're passing OffsetDateTime instead of
ZonedDateTime into the PostgreSQL JDBC driver, the "to string"
serialization seems to be different. It's however logically equivalent,
so I will keep it as is and just update the tests.

Change-Id: I6f2f973669d6ef5b6fcb3808e9793d7bcdc2c3c1
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/2684
Reviewed-by: Alex Sánchez <alex.sanchez@streamsets.com>