SDC-16076. Implement test for FieldMapper to add error records on invalid field path
We test that if an expression evaluates to an invalid value
e.g. a field path containing a quote, or cannot be evaluated at runtime,
a record error is added instead of a stage error as it was before.

If an expression evaluates to a valid value the field is
modified as needed.

The tests cover aggregation_expression, conditional_expression and
mapping_expression configurations.

Change-Id: I880e2d899b7a43be956e7bd09c972678487449ad
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/2725
Reviewed-by: Santhosh Kumar Manavasi Lakshminarayanan <santhosh@streamsets.com>