SDC-16108. Implement tests to verify a fix SQS pipeline data loss on pipeline stop
In this test we want to make sure that the SQS consumer supports
the At Least Once and At Most Once modes. If we receive a message but
the record doesn't reach the destination for the At Least Once mode we
expect the message stays in the queue.
For the At Most Once mode we expect the message gets deleted from the
queue even if a record doesn't reach a destination.
To make sure a record doesn't reach a destination we put a JavaScript
evaluator which always throws an error. This makes the pipeline to stop
with the RUN_ERROR status.

Change-Id: I58b1828ccd972ba6f04f0e4bde62800268246995
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/2740
Reviewed-by: Santhosh Kumar Manavasi Lakshminarayanan <santhosh@streamsets.com>
Tested-by: <streamsets-ci@streamsets.com>