SDC-16154: PostgreSQL Interval Types are failing after JDBC Driver update
It's actually kind of expected - the driver is the piece of code that
serializes the weird types into String and with every upgrade it's
expected that the format might slightly change - provided the logical
value remains the same. In this case they stopped emiting extra zero for
seconds which also seems fair.

Change-Id: Ia1d47a0f5beb7df5a5e818161c25f4b5f46a415c
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/2754
Reviewed-by: Alex Sánchez <alex.sanchez@streamsets.com>