SDC-16193. Implement test for mutlitable JDBC produces duplicate records with SWITCH_TABLE
We want to make sure that there are no duplicates or data loses when
the number of threads is less than the number of tables; is equal to
the number of tables or is greater than the number of tables. And we
want to run the same test for all batch strategies with partitioning
enabled and disabled too.

To run the test we first create 3 tables and populate them with
12000, 36000 and 6000 random records respectively.

To prove the connector works as expected we will
1) test that the number of records written is as expected for very
table, not more (to exclude data duplicates) and not less
(to exclude data loses).
2) test that at the end we get all expected events: 3 table-finished
events (1 for each table), 1 schema-finished event with 3 table names
and 1 no-more-data event.

Change-Id: Ic560055585c65c16545ab0f0c8b0b62d57e78080
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/2768
Tested-by: <streamsets-ci@streamsets.com>
Reviewed-by: Dima Spivak <dima@streamsets.com>