SDC-16200. Implement tests for JSON Parser doesn't handle byte array fields properly
By definition the JSON parser processor expects a serialized JSON
object in a string field.

Depending on a field type we expect:
1. JSON parser to parse a field value when a field type supports
conversion to a string and the converted value is a valid JSON.
2. A JSONP_XXX error record to be added when a field type supports
conversion to a string and the converted value is a not valid JSON.
3. An API_XXX error record to be added when a field type doesn't
support conversion to a string.

Change-Id: I6643629d9e8bb1dab0c61121363cbb18c120b618
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/2776
Tested-by: <streamsets-ci@streamsets.com>
Reviewed-by: Santhosh Kumar Manavasi Lakshminarayanan <santhosh@streamsets.com>