SDC-16247. Implement tests for Cache doesn't work in Kudu Lookup Processor
We want to test that when the caching is enabled we get obsolete data.
So we will insert a record in the storage and then we will create
a file. The pipeline reads the file, then loads a record from the
storage and replaces the value we have read from the file with the one
we have read from the storage.

If we update the record in the storage and then create the second file
on the file system then due to the caching enabled the value in the
file will be replaced with the storage value we have loaded before and
the updated record will not be loaded.

If the caching is disabled the new value will be used.

Seems there is a bug in the current implementation of Kudu Lookup
Processor and the caching doesn't work, thus this test fails.
If the issue we identified is really a bug then after fixing it the
test will pass as expected.

Please review the test and confirm that the scenario is valid and the
test must pass if the cache implementation is not broken.

Change-Id: Ibe58292b02ba68a3c7bc369749a8a85a9d6440e4
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/2793
Reviewed-by: Alex Sánchez <alex.sanchez@streamsets.com>
Reviewed-by: Hugo Prol <hugo@streamsets.com>