SDC-16235. Implement tests for Kudu Lookup Processor Retry on Missing Value
Normally if a value in a table is missing and a default value is
specified, the default value is cached when local caching is enabled.
We want to test that if the retry_on_missing_value field configuration
is on, the default value is not cached and if next time a value
appears in the table, that value is used even if the cache is enabled.

Change-Id: I44ba451014e9dd9840eb0eff92c48954595283bf
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/2794
Tested-by: <streamsets-ci@streamsets.com>
Reviewed-by: Hugo Prol <hugo@streamsets.com>