SDC-16355: Improve error cases for Kafka cluster pipelines
What is happening right now on CDP 7 is that the Kafka cluster tests
don't work at all - the pipeline doesn't even start, so trying to stop
it leads to an error that then skips stopping the second pipeline --
that is left over and running on the node taking resources on a stressed
machine.

This patch ensures that the finally block stops both pipeline no matter
the errors in the first pipeline stop.

Change-Id: I9f981e03fc870e68f9bb54ac758c2848a8a59fb3
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/2935
Reviewed-by: Alex Sánchez <alex.sanchez@streamsets.com>