SDC-16448: Kafka standard tests sometimes time-out
Bear with me as this will be slightly longer explanation for one line
change.

I was looking into the various Kafka failures and noticed a few things:

1) They are pretty much isolated to really old Kafka versions (before
   2.x)

2) 14 tests requires 10+ minutes to execute even after using already spawn
   up Data Collector instances. That is almost a minute per test!

3) The execution actually varies, sometimes it's 10 minutes, sometimes
   it's almost 12 minutes.

Thus it seems to me that for some reason the older Kafka versions are
slower for the test use case and sometimes time out after the default 30
second wait time. I've tried to increase the timeout to 120 seconds and
with this change all the flakiness went away on my box - and the
execution remained in the 10-13 minute range.

I would like to get this change merged and have it running for a few
days to see if that helps on Jenkins too. And if so, I will propagate it
to the other relevant tests.

Change-Id: I25dc8d445d8edb7933d6b0a7744b9ce1090bb494
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/3007
Reviewed-by: Martin Balzamo <martin@streamsets.com>