SDC-16453: Set explicit consumer groups in Kafka tests
This is another attempt to make the tests less flaky similarly to
SDC-16448 - the thesis now is that by default we're using the same
consumer group in all tests and based on subjective observations it
seems that sometimes the older Kafka versions take more time to get new
consumer in the same group to start reading a new topic, so the idea is
to try setting explicit consumer group and see if that helps.

It is a bit of a shot in the dark, but I have seen a bit lesser spread
of runtime on my box with this patch (much more closer to the 10-11
minutes then going to 12 and 13).

Change-Id: I6fdc17a92c8744f2fd245656e84eb8ecde7a09dd
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/3009
Reviewed-by: Martin Balzamo <martin@streamsets.com>