SDC-16458: Attempt to retry submitting failed cluster mode pipelines
This is a continuation of my previous tweak done in SDC-16409. With that
output I was able to reproduce the error a few times times. Basically
what seems to be happening is that we submit a cluster job - the job is
populated by Spark properly (spark uploads all the jars into staging
directory), but YARN won't accept it. Why? That I was not able to figure
out in the logs at all :-( Our pipeline in SDC actually fails with
CONNECT_ERROR very clearly, but the test isn't written to detect it and
hence the usual weird flakiness.

I did find a correlation on my machine though - if i'm running Nest cam
(yeah checking if kiddo is sleeping right), Zoom, music, IntelliJ - and
thus I'm stressing my local box, I'm able to repro the same failure at
least sometimes, while when my box doesn't do anything except of running
the tests, I can't reproduce at all. This leads me to believe that when
the machine is stressed, something somewhere between Spark <> Yarn
decided to not accept the job and hence the failure.

Considering that this is something on the cluster and the state of
cluster mode, I kind of gave in and decided to do 'workaround' on the
test side. And rather detect when the job wasn't submitted and re-submit
it again rather then spending further time investigating where exactly
is the job rejected.

Change-Id: I9b9099b98343f2039a1233d41fc6e2509d4167aa
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/3015
Reviewed-by: Alex Sánchez <alex.sanchez@streamsets.com>