SDC-16532: Update S3 bucket name constant for "minsize"
It seems that someone have claimed the 1a1 bucket, so I've slightly
updated it - probably still not great (and final solution), but will do
for now.

I've also updated the error handling - so that the exception propagated
from the test (and to bifocals) is the right now and not something
failing in finally clean up.

Change-Id: Ia6ae425444ba462ee527f5e12ea4152f091bae8a
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/3070
Reviewed-by: Alex Sánchez <alex.sanchez@streamsets.com>