SDC-16553: Randomize minsize bucket name for S3 standard tests
The new constant was nice in the sense that it wasn't pre-existing
bucket, but when I run all upgrade tests at the same time, they attempt
to create and use the same bucket name which means - another flaky
tests.

This patch tries slightly different approach - since the bucket names
exist in a shared space, their definition in the test "header" is a
lamdba function and then we have retry logic that will generate up to 10
different buckets and try to "grab" the first existing one.

Change-Id: Ie112743adf0803dca7f97f75312647b5c330be30
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/3086
Reviewed-by: Alex Sánchez <alex.sanchez@streamsets.com>