SDC-16745: Add simple standard test coverage reporting script
Simple script that generates "coverage" report for the standard tests -
it's meant as a static analysis tool, not really running and fetching
report from Bifocals. It's also not super smart, but it's good enough to
get basic idea of what we have and what we're missing.

I'm fully expecting that we will eventually get rid of the script and
incorporate the coverage into Bifocals - when the time will be right.

Change-Id: Iedc99986d2bae8e4bf3b44255200b52bc182a3f0
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/3156
Reviewed-by: Alex Sánchez <alex.sanchez@streamsets.com>