COLLECTOR-494. Test TimeEL on Hive Query Executor.
I added more tests to test_hive_query_executor.py for SDCs of version
4.4.0 or higher, which include the TimeEL feature. The tests check that a
pipeline with a Hive Query Executor query that contains a TimeEL condition
can be created and executed without errors, and that the timestamps stored
when running the pipeline with TimeEL are correct.
It also checks that a Hive Query Executor query with a TimeEL condition
works as expected when different timezones are set in the pipeline
configuration.

Upgrade tests for the Hive Query Executor after the addition of TimeEL are
included in I3fe541da15660dd5f03951e3cde33bb7b2d38269.

Change-Id: I63b4a2545d3162ca809c40e28265e48c3f2d180a
Reviewed-on: https://review.streamsets.dev/c/datacollector-tests/+/3399
Reviewed-by: Santhosh Kumar Manavasi Lakshminarayanan <santhosh@streamsets.com>
Reviewed-by: Alex Sánchez <alex.sanchez@streamsets.com>