COLLECTOR-1959. JDK Max memory test failed when executing with Java 17.
It was failing because it was expecting severity RED but was returning
SEVERITY YELLOW that is better. The reason is because java 17 is
reporting exactly 1GB that is considered better thant the 989MB reported
when executing with Java 8. The test has changed to that it only fails
when the reported severity is worst than expected but not when it is
better.

Change-Id: I8677f5ef2206c5b745f1da44c0f13da556a550ee
Reviewed-on: https://review.streamsets.net/c/datacollector-tests/+/58382
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Joaquín Bo Díaz <joaquin.bo.diaz@streamsets.com>