COLLECTOR-2981. stage.test_azure_blob_storage_origin.test_file_postprocessing[action=ARCHIVE] failing for SHARED_KEY auth
It looks like the SHARED_KEY ADLS GEn2 environment is not prepared to allow copies of the blobs, so we are getting a 401 unauthorized error when trying so. Given that this test is passing for OAuth ADLS Gen2 environment, I think we can skip the test for the unprepared environment (basically some permissions needed if doing this operation with SHARED_KEY auth). If it gives problems in the future, we can un-skip the test and revisit the issue.

Change-Id: Ie040fe347628eb6df94a9967c456fa35e3e2102f
Reviewed-on: https://review.streamsets.net/c/datacollector-tests/+/60255
Reviewed-by: Mikel Solabarrieta Román <mikelsr@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>