COLLECTOR-4247. Remove MonthlyPlatformEventsUsageEntitlement from limits info to be gathered
We eventually stopped receiving this information. EP is investigating
whether it is due any expired licence, as this value seems to be
provided when "Org With Add-On License: Monthly Usage-Based Entitlement"
as per Salesforce documentation. In the meanwile as we are only using
this info just to log it and the tests seems to be passing, I'm commenting
this line out.

Change-Id: I866c9cb16a9f4e0278b8c5dcf3f21b65513f7e84
Reviewed-on: https://review.streamsets.net/c/datacollector-tests/+/66229
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Neeraj Jain <neeraj@streamsets.com>